pyproject.toml
[tool.poetry]
name = "veda-optimized-data-pipelines"
version = "0.1.0"
description = "Cloud pipelines for ingesting cloud optimized data"
authors = ["IMPACT", "DevSeed"]
license = "MIT"
packages = [
    { include = "scripts" }
]

[tool.poetry.dependencies]
python = "^3.8"
argparse = "^1.4.0"
boto3 = "^1.21.43"
moto = {extras = ["s3"], version = "^3.1.16"}
pytest = "^7.1.2"
pydantic = "^1.9.1"
pystac = "1.4.0"
python-cmr = "0.7.0"
rasterio = "1.3.0"
rio-stac = "0.4.2"
shapely = "1.8.4"
smart-open = "4.2.0"
exceptiongroup = "1.0.0rc9"
psycopg = "3.1"
psycopg-pool = "3.1.1"
pypgstac = "0.6.6"

[tool.poetry.dev-dependencies]
"aws_cdk.core" = "^1.151.0"
"aws_cdk.aws_iam" = "^1.151.0"
"aws_cdk.custom_resources" = "^1.151.0"
"aws_cdk.aws_stepfunctions" = "^1.151.0"
"aws_cdk.aws_events" = "^1.151.0"
"aws_cdk.aws_events_targets" = "^1.151.0"
"aws_cdk.aws_secretsmanager" = "^1.151.0"
"aws_cdk.aws_lambda" = "^1.151.0"
"aws_cdk.aws_stepfunctions_tasks" = "^1.151.0"
"aws_cdk.aws_ec2" = "^1.151.0"
"aws_cdk.aws_sqs" = "^1.151.0"
"aws_cdk.aws_s3" = "^1.151.0"
"aws_cdk.aws_lambda_event_sources" = "^1.151.0"
"aws_cdk.aws_lambda_python" = "^1.151.0"
boto3-stubs = {extras = ["s3"], version = "^1.24.45"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
insert-collection = "scripts.collection:insert"
delete-collection = "scripts.collection:delete"
update-collection = "scripts.collection:update"
insert-item = "scripts.item:insert"
delete-item = "scripts.item:delete"
update-item = "scripts.item:update"
deploy = "scripts.cdk:deploy"
destroy = "scripts.cdk:destroy"
